+++ R/EffectSizePropStat.R
@@ -30,7 +30,7 @@ EffectSizePropStat = function(sample.list, parameter) {
   }
 
   else if (call == TRUE) {
-    result = list("Effect size")
+    result = list("Effect size (proportion)")
   }
 
   return(result)