+++ R/RatioEffectSizeEventStat.R
@@ -13,6 +13,14 @@ RatioEffectSizeEventStat = function(sample.list, parameter) {
     if (length(sample.list)!=4)
       stop("Analysis model: Four samples must be specified in the RatioEffectSizeEventStat statistic.")
 
+    if (is.na(parameter[[2]])) method = "Log-Rank"
+    else {
+      if (!(parameter[[2]]$method %in% c("Log-Rank", "Cox")))
+        stop("Analysis model: HazardRatioStat statistic : the method must be Log-Rank or Cox.")
+
+      method = parameter[[2]]$method
+    }
+
     result1 = EffectSizeEventStat(list(sample.list[[1]], sample.list[[2]]), parameter)
 
     result2 = EffectSizeEventStat(list(sample.list[[3]], sample.list[[4]]), parameter)
@@ -22,7 +30,10 @@ RatioEffectSizeEventStat = function(sample.list, parameter) {
   }
 
   else if (call == TRUE) {
-    result = list("Ratio of effect size")
+    if (is.na(parameter[[2]])) result = list("Ratio of effect size (event)")
+    else {
+      result = list("Ratio of effect size (event)", "method = ")
+    }
   }
 
   return(result)