+++ R/RatioEffectSizePropStat.R
@@ -45,7 +45,7 @@ RatioEffectSizePropStat = function(sample.list, parameter) {
   }
 
   else if (call == TRUE) {
-    result = list("Ratio of effect size")
+    result = list("Ratio of effect size (proportion)")
   }
 
   return(result)